Refactor XQG customization variables
This is to make it easier to modify the format of the XQuery to enable
things such as using recursion instead of a for loop to find the root
apply node.

Some unit tests were also changed to reflect the fact that the user
shouldn't be able to specify random headers/footers - the user should
only need to specify the namespace, the root element, and the return
format.